+++ authentication-server/src/controllers/AuthController.ts
@@ -2,6 +2,7 @@ import type { RequestHandler } from "express";
 import { errorHandling, strictCheck } from "../utils/ApplicationError.js";
 import type AuthService from "../services/AuthService.js";
 import DeviceDetector from "device-detector-js";
+import type { Hardware } from "../types/dbSchema.js";
 
 export default class AuthController {
     private authService: AuthService;
@@ -57,6 +58,72 @@ export default class AuthController {
     }
 
     /**
+     * Handles get session Request
+     * 
+     * Inorder Flow: 
+     * - Retrieve sessionId and Entity from request
+     * - Retrieve session data
+     * - structure and send response
+     */
+    getSessions: RequestHandler = async (req, res) => {
+        try {
+            const sessionId = strictCheck(req.sessionId, 500, "Internal Server Error");
+            const entity = strictCheck(req.user ?? req.organization, 500, "Internal Server Error");
+
+            const sessionData = await this.authService.getSessions(sessionId, entity);
+
+            res.status(200).json({
+                success: true,
+                message: "Sessions have been retrieved successfully",
+                data:sessionData ?? []
+            });
+        } catch (error) {errorHandling(error, res);}
+    }
+
+    /**
+     * Handles revoke session Request
+     * 
+     * Inorder Flow: 
+     * - Retrieve maskedId from request body
+     * - Revoke session by maskedId
+     * - structure and send response
+     */
+    revokeSession: RequestHandler = async (req, res) => {
+        try {
+            const { _id } = req.body;
+
+            await this.authService.revokeSession(_id);
+
+            res.status(200).json({
+                success: true,
+                message: "Session has been revoked successfully"
+            });
+        } catch (error) {errorHandling(error, res);}
+    }
+
+    /**
+     * Handles revoke all sessions Request
+     * 
+     * Inorder Flow: 
+     * - Retrieve sessionId, entity from request
+     * - Revoke all sessions
+     * - structure and send response
+     */
+    revokeAllSessions: RequestHandler = async (req, res) => {
+        try {
+            const sessionId = strictCheck(req.sessionId, 500, "Internal Server Error");
+            const entity = strictCheck(req.user ?? req.organization, 500, "Internal Server Error");
+
+            await this.authService.revokeAllSessions(sessionId, entity.id);
+
+            res.status(200).json({
+                success: true,
+                message: "Sessions have been revoked successfully"
+            });
+        } catch (error) {errorHandling(error, res);}
+    }
+
+    /**
      * Handles user login Request
      * 
      * Inorder Flow:
@@ -70,7 +137,7 @@ export default class AuthController {
             const { email, password }: {email:string, password:string } = req.body;
             const ip = req.ip as string;
             const userAgent = req.headers["user-agent"] as string;
-            const hardware = this.deviceDetector.parse(userAgent);
+            const hardware = this.deviceDetector.parse(userAgent) as Hardware;
 
             const loginPayload = strictCheck(await this.authService.userLogin({email, password, ip, hardware}), 401, "Invalid Credentials");
 
@@ -106,7 +173,7 @@ export default class AuthController {
             const { email, password }: {email:string, password:string } = req.body;
             const ip = req.ip as string;
             const userAgent = req.headers["user-agent"] as string;
-            const hardware = this.deviceDetector.parse(userAgent);
+            const hardware = this.deviceDetector.parse(userAgent) as Hardware;
 
             const loginPayload = strictCheck(await this.authService.organizationLogin({email, password, ip, hardware}), 401, "Invalid Credentials");