+++ frontend-server/src/contexts/ApplicationContext.tsx
@@ -1,6 +1,6 @@
 import { API_ROUTES } from "@/config/routes";
 import useApiService from "@/hooks/useApiService";
-import type { Client } from "@/types";
+import type { Client, Session } from "@/types";
 import type { ProfileSchema, CreateClientSchema } from "@/types/validation";
 import { createContext, useContext, type ReactNode } from "react";
 import { useNotification } from "./NotificationContext";
@@ -10,12 +10,15 @@ interface ApplicationContextType {
 	updateProfile: (values: ProfileSchema) => Promise<boolean>;
 	clients: () => Promise<Client[] | void>;
 	createClient: (values: CreateClientSchema) => Promise<Client | null>;
+	sessions: () => Promise<Session[] | void>;
+	revokeSession: (_id: string) => Promise<boolean>;
+	revokeAllSessions: () => Promise<boolean>;
 }
 
 const ApplicationContext = createContext<ApplicationContextType | undefined>(undefined);
 
 export function ApplicationProvider({ children }: { children: ReactNode }) {
-	const { getRequest, postRequest, putRequest} = useApiService(import.meta.env.VITE_API_URL);
+	const { getRequest, postRequest, putRequest, deleteRequest} = useApiService(import.meta.env.VITE_API_URL);
 	const { error, success } = useNotification();
 
 	async function profile(): Promise<ProfileSchema | void>{
@@ -72,12 +75,58 @@ export function ApplicationProvider({ children }: { children: ReactNode }) {
 			return null;
 		};
 
+	async function sessions(): Promise<Session[] | void>{
+		const response = await getRequest <Session[]>(API_ROUTES.session);
+		if(!response) return;
+
+		if (!response.success && response.error) {
+			error("Unable to load sessions", response.error);
+			return;
+		}
+
+		return response.data;
+	}
+
+	async function revokeSession(_id: string): Promise<boolean> {
+		const response = await deleteRequest <undefined>(API_ROUTES.revokeSession,{_id: _id});
+		if (!response) return false;
+
+		if (response.success && response.message) {
+			success("Session revoked", response.message);
+			return true;
+		}
+
+		if (!response.success && response.error){
+			error("Unable to revoke session", response.error );
+			return false;
+		}
+		return false;
+	}
+
+	async function revokeAllSessions(): Promise<boolean> {
+		const response = await deleteRequest <undefined>(API_ROUTES.revokeAllSessions,{});
+		if (!response) return false;
+
+		if (response.success && response.message) {
+			success("All Sessions are revoked", response.message);
+			return true;
+		}
+
+		if (!response.success && response.error){
+			error("Unable to revoke sessions", response.error );
+			return false;
+		}
+		return false;
+	}
 	
 	return <ApplicationContext.Provider value={{
 		profile,
 		updateProfile,
 		clients,
-		createClient
+		createClient,
+		sessions,
+		revokeSession,
+		revokeAllSessions
 	}}>{children}</ApplicationContext.Provider>;
 }