+++ frontend-server/src/types/index.ts
@@ -18,3 +18,29 @@ export interface Client {
 	responseType: "code";
 	scopes: string[];
 }
+
+interface Hardware {
+	client: {
+		name?: string;
+		version?: string;
+		type?: string;
+	};
+	os: {
+		name?: string;
+		version?: string;
+	};
+	device: {
+		type?: string;
+		vendor?: string;
+		model?: string;
+	};
+}
+
+export interface Session {
+	_id: string;
+	hardware: Hardware;
+	createdAt: Date;
+	expiresAt: Date;
+	ip: string;
+	isCurrent: boolean;
+}