+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The Following rules must be followed at all times
 ### Added
 - Added initial release documentation with proper structure ( `CHANGELOG.md` and `RELEASE_NOTES.md`)
 - Added Email Verification feature
+- Added Session management feature
 
 ### Changed
 - Improvised the UI for all Pages